+++ dash_extensions/enrich.py
@@ -30,6 +30,9 @@ from dash import (  # lgtm [py/unused-import]
     callback_context,
     callback,
     clientside_callback,
+    page_container,
+    page_registry,
+    register_page
 )
 from dash._utils import patch_collections_abc
 from dash.dependencies import _Wildcard, DashDependency  # lgtm [py/unused-import]
@@ -38,7 +41,7 @@ from flask import session
 from flask_caching.backends import FileSystemCache, RedisCache
 from more_itertools import flatten
 from collections import defaultdict
-from typing import Dict, Callable, List, Union, Any, Tuple
+from typing import Dict, Callable, List, Union, Any, Tuple, Optional
 from datetime import datetime
 from dash_extensions import CycleBreaker
 
@@ -989,9 +992,9 @@ class MultiplexerTransform(DashTransform):
             mp_id = _mp_id(output, i)
             mp_id_escaped = _escape_wildcards(mp_id)
             # Create proxy element.
-            proxies.append(_mp_element(mp_id_escaped))
+            proxies.append(_mp_element(mp_id_escaped.copy()))
             # Assign proxy element as output.
-            callback.outputs[callback.outputs.index(output)] = Output(mp_id_escaped, _mp_prop())
+            callback.outputs[callback.outputs.index(output)] = Output(mp_id_escaped.copy(), _mp_prop())
             # Create proxy input.
             inputs.append(Input(mp_id, _mp_prop()))
         # Collect proxy elements to add to layout.
@@ -1278,6 +1281,253 @@ class NoOutputTransform(DashTransform):
 
 # endregion
 
+# region [Magic] transform
+
+class OperatorOutput(Output):
+    """
+    Like a normal Output, but enables list and dict manipulations.
+    """
+
+
+class Operator:
+    def __init__(self, path: Optional[List] = None, operations: Optional[List] = None):
+        self.path = path if path is not None else []
+        self.operations = operations if operations is not None else []
+
+    def __getitem__(self, key):
+        return Operator(self.path + [key], operations=self.operations)
+
+    def __setitem__(self, key, item):
+        self.path += [key]
+        return self.assign(item)
+
+    def collect(self, opr, **kwargs):
+        self.operations.append(dict(opr=opr, pth=self.path, **kwargs))
+        self.path = []
+        return self
+
+    @property
+    def list(self):
+        return ListOperator(self)
+
+    @property
+    def dict(self):
+        return DictOperator(self)
+
+    def assign(self, item):
+        return self.collect("assign", item=item)
+
+    def apply(self):
+        return self.operations
+
+
+class ListOperator:
+    def __init__(self, operator: Optional[Operator] = None):
+        self.operator = operator
+
+    def _collect(self, opr, **kwargs):
+        return self.operator.collect(f"list_{opr}", **kwargs)
+
+    def apply(self):
+        return self.operator.apply()
+
+    def append(self, item):
+        return self._collect("append", item=item)
+
+    def extend(self, iterable):
+        return self._collect("extend", array=list(iterable))
+
+    def insert(self, index, item):
+        return self._collect("insert", item=item, index=index)
+
+    def remove(self, item):
+        # NB: Remove ALL occurances of item, not just the first one.
+        return self._collect("remove", item=item)
+
+    def pop(self, index):
+        return self._collect("pop", index=index)
+
+    def clear(self):
+        return self._collect("clear")
+
+    def sort(self):
+        return self._collect("sort")
+
+    def reverse(self):
+        return self._collect("reverse")
+
+
+class DictOperator:
+    def __init__(self, operator: Optional[Operator] = None):
+        self.operator = operator
+
+    def _collect(self, opr, **kwargs):
+        return self.operator.collect(f"dict_{opr}", **kwargs)
+
+    def apply(self):
+        return self.operator.apply()
+
+    def set(self, key, item):
+        return self._collect("set", key=key, item=item)
+
+    def pop(self, key):
+        return self._collect("pop", key=key)
+
+    def update(self, obj):
+        return self._collect("update", obj=obj)
+
+    def clear(self):
+        return self._collect("clear")
+
+
+class OperatorTransform(DashTransform):
+    def __init__(self):
+        super().__init__()
+        self.components = []
+        self.operator_outputs = []
+        self.blueprint = DashBlueprint()
+
+    def transform_layout(self, layout):
+        children = _as_list(layout.children) + self.components
+        layout.children = children
+
+    def _apply(self, callback, output):
+        original_id = output.component_id
+        relay_id = _relay_id(original_id)
+        if str(output) not in self.operator_outputs:
+            # Append new relay component.
+            relay_component = dcc.Store(id=relay_id)
+            self.components.append(relay_component)
+            # Add clientside callback to perform modifications.
+            self.blueprint.clientside_callback(f"""function(operations, current){{
+                // Handle empty init call.
+                if (typeof operations === 'undefined'){{
+                    return window.dash_clientside.no_update;
+                }}
+                // Map non-list actions to list to enable iteration.
+                if (!(Array.isArray(operations))){{
+                    operations = [operations];
+                }}
+                // Function for resolving sub elements.
+                const drill = (obj, pth, lvl=0) => {{
+                    if(lvl === pth.length){{return obj;}}
+                    return drill(obj[pth[lvl]], pth, lvl+1);
+                }}
+                // Wrap current in list to enable index access.
+                lst = [current];
+                // Action.
+                for (const x of operations) {{
+                    let pth = [0].concat(x.pth);
+                    let idx = pth[pth.length - 1];
+                    let obj = drill(lst, pth.slice(0,-1));
+                    switch(x.opr) {{
+                      case "assign":
+                        obj[idx] = x.item;             
+                        break;
+                      // List action(s).
+                      case "list_append":
+                        obj[idx].push(x.item)
+                        break;
+                      case "list_extend":
+                        obj[idx] = obj[idx].concat(x.array);
+                        break;
+                      case "list_insert":
+                        obj[idx].splice(x.index, 0, x.item);
+                        break;
+                      case "list_remove":
+                        obj[idx] = obj[idx].filter(function(ele){{
+                            return ele != x.item;
+                        }});
+                        break;
+                      case "list_pop":
+                        obj[idx].splice(x.index, 1);
+                        break;
+                      case "list_reverse":
+                        obj[idx].reverse();
+                        break;
+                      case "list_sort":
+                        // TODO: Make it possible to inject sorting function
+                        obj[idx].sort();
+                        break;
+                      case "list_clear":
+                        obj[idx] = []             
+                        break;
+                      // Dict action(s).
+                      case "dict_set":
+                        obj[idx][x.key] = x.item;
+                        break;
+                      case "dict_pop":
+                        delete obj[idx][x.key];
+                        break;
+                      case "dict_clear":
+                        obj[idx] = {{}};             
+                        break;
+                      case "dict_update":
+                        obj[idx] = {{
+                            ...obj[idx],
+                            ...x.obj
+                        }};
+                      // Unknown action(s).
+                      default:
+                        console.log("Received unknown action for component {original_id}.");
+                        console.log(x);
+                        console.log("Update will be skipped.");
+                    }}
+                }}
+                // Make sure out ref != input ref (otherwise, React can get confused)
+                const result = lst[0];
+                if(Array.isArray(result)){{
+                    return Array.from(result)
+                }}
+                if(typeof result === 'object'){{
+                    return Object.assign({{}}, result);
+                }}
+                return result;
+            }}""", output, Input(relay_id, "data"), State(output.component_id, output.component_property))
+            # Record binding.
+            self.operator_outputs.append(str(output))
+        # Modify callback in-place to route output to the relay.
+        callback.outputs[callback.outputs.index(output)] = Output(relay_id, "data")
+        # Run apply if needed.
+        f = callback.f
+        callback.f = apply_operator()(f)
+
+    def apply_serverside(self, callbacks):
+        for callback in callbacks:
+            for output in callback.outputs:
+                if isinstance(output, OperatorOutput):
+                    self._apply(callback, output)
+        return callbacks
+
+    def apply_clientside(self, callbacks):
+        return callbacks + self.blueprint.clientside_callbacks
+
+    def get_dependent_transforms(self):
+        return [MultiplexerTransform()]
+
+
+def apply_operator():
+    def wrapper(f):
+        @functools.wraps(f)
+        def decorated_function(*args):
+            output = f(*args)
+            if isinstance(output, Operator):
+                return output.apply()
+            if hasattr(output, "__len__"):
+                output = [o.apply() if isinstance(o, Operator) else o for o in output]
+            return output
+
+        return decorated_function
+
+    return wrapper
+
+
+def _relay_id(uid):
+    return f"{uid}_operator_relay"
+
+
+# endregion
+
 # region Batteries included dash proxy object
 
 
@@ -1292,6 +1542,7 @@ class Dash(DashProxy):
             CycleBreakerTransform(),
             BlockingCallbackTransform(),
             ServersideOutputTransform(**output_defaults),
+            OperatorTransform()
         ]
         super().__init__(*args, transforms=transforms, **kwargs)